requirements.txt
flask
openai
line-bot-sdk
faiss-cpu
sentence-transformers
scikit-learn
pandas